Add observersss to replace text in added/modified nodesss
A number of people have commented saying the extension doesn't
replace *all* text and this is partly because dynamic content isn't
handled. Right now the extension simply replaces text after the page
is loaded, so we don't handle anything that's not present at that
point.

To help fix this, I've added observers that watch for modifications
in document.body and document.title and replace text appropriately.