Reduce replace() calls for gendered/aged snake people
This commit obviates the need for so many replace() calls by simply
improving the regexes and using capture groups in a crafty way. These
modifications do the same replacements as before while also taking
care of some cases that weren't being handled.

In the end, this uses about a quarter of the number of replace()
calls that were originally in place here.