Update source to work better with lazy-loading sites
A site such as Facebook, which has infinite scroll, will not change new instances of "the cloud" when the user scrolls down and more posts are added to the newsfeed. By adding an event listener, we can also target such instances and thus get even closer to the dream of true 100% aerosol-rectal conversion.